chore: moves files into folders depending on type
