+++ src/core_modules/capture-core/rules/engine/commonUtils/dateUtils.js
@@ -1,6 +1,6 @@
 // @flow
 import moment from 'moment';
-import trimQuotes from '../commonUtils/trimQuotes';
+import trimQuotes from './trimQuotes';
 import type { IMomentConverter } from '../rulesEngine.types';
 
 export default function getDateUtils(converterObject: IMomentConverter) {
