+++ src/core_modules/capture-core/rules/engine/helpers/OptionSetHelper.js
