+++ src/core_modules/capture-core/rules/engine/index.js
@@ -1,5 +1,4 @@
 // @flow
 export { default as RulesEngine } from './RulesEngine';
 export { effectActionsConstants } from './effectActions.const';
-export { default as processTypes } from './rulesEffectsProcessor/processTypes.const';
 export * from './rulesEngine.types';
