+++ src/core_modules/capture-core/rules/engine/processors/ValueProcessor.js
