+++ src/core_modules/capture-core/rules/engine/processors/rulesEffectsProcessor/processTypes.const.js
