+++ src/core_modules/capture-core/rules/engine/services/VariableService/variablePrefixes.const.js
